fix(upload): 修复编译错误——totalLen 改为 int，避免 long→int lossy conversion（此前导致 CI 构建失败、进度功能无 APK 可装）

+++ app/src/main/java/com/pan/mobile/MainActivity.java
@@ -1248,7 +1248,7 @@ public class MainActivity extends Activity {
                         put.setDoOutput(true);
                         put.setFixedLengthStreamingMode(all.length);
                         java.io.OutputStream pos = put.getOutputStream();
-                        final long totalLen = all.length;
+                        final int totalLen = all.length; // byte[] 长度即 int，与 Math.min/setFixedLengthStreamingMode 保持一致，避免 long→int lossy
                         final int upChunk = 262144; // 256KB，兼顾真实进度反馈与无谓回调开销
                         int sent = 0;
                         while (sent < totalLen) {
